requirements.txt
python-dotenv
flask
Flask[async]
azure-storage-blob